lce id's int data type with long, added extra test scenarious for addition of second object, shortened the PersistentUserRepositoryTest getUsersList scenario.

+++ schema/user_info_ddl.sql
@@ -1,18 +1,18 @@
 create TABLE address (
-    id integer not null,
-    user_id integer not null,
+    id BIGINT not null,
+    user_id BIGINT not null,
     address text not null,
     PRIMARY KEY (id),
     FOREIGN KEY (user_id) REFERENCES users (id)
 );
 create TABLE contact (
-    id integer not null,
-    user_id integer not null,
+    id BIGINT not null,
+    user_id BIGINT not null,
     phone_number text not null,
     PRIMARY KEY (id),
     FOREIGN KEY (user_id) REFERENCES users (id)
 );
 create TABLE users (
-    id integer PRIMARY KEY not null,
+    id BIGINT PRIMARY KEY not null,
     name text not null
 );
